Move model definitions to core-model, domains-model
Move model definitions to core-model, domains-model

Closes #517
