+++ v0.4/platform/core-model/.eslintrc.js
@@ -1,0 +1,1 @@
+../../.eslintrc.js.template
