+++ v0.4/platform/core-model/package.json
@@ -1,35 +1,32 @@
 {
-  "name": "@anticrm/model",
+  "name": "@anticrm/core-model",
   "version": "0.2.0",
   "main": "src/index.ts",
   "author": "Anticrm Platform Contributors (https://github.com/anticrm)",
   "license": "EPL-2.0",
   "scripts": {
     "build": "tsc",
-    "build:docs": "api-extractor run --local && copyfiles -u 1 ./_api-extractor-temp/doc-models/* ../../_api-extractor-temp/",
-    "test": "echo 'no tests'",
+    "build:docs": "echo 'noapi'",
     "lint": "eslint --max-warnings=0 src",
-    "lint:fix": "eslint --fix src"
-  },
-  "dependencies": {
-    "@anticrm/core": "0.2.0",
-    "@anticrm/status": "0.4.2",
-    "lodash": "^4.17.21",
-    "reflect-metadata": "^0.1.13"
+    "lint:fix": "eslint --fix src",
+    "test": "echo 'no-tests",
+    "genmodel": "ts-node src/__tests__/genmodel.ts"
   },
   "devDependencies": {
-    "@microsoft/api-extractor": "^7.14.0",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
-    "eslint": "^7.25.0",
+    "eslint": "^7.24.0",
     "eslint-config-standard": "^16.0.2",
     "eslint-config-standard-with-typescript": "^20.0.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.1.0",
-    "simplytyped": "^3.3.0",
-    "ts-node": "^9.1.1",
     "typescript": "^4.2.4",
     "@types/node": "~15.0.2"
+  },
+  "dependencies": {
+    "@anticrm/status": "~0.4.2",
+    "@anticrm/core": "^0.2.0",
+    "@anticrm/model": "^0.2.0"
   }
 }
