+++ v0.4/platform/domains-model/.eslintrc.js
@@ -1,0 +1,1 @@
+../../.eslintrc.js.template
