+++ v0.4/platform/domains-model/src/__tests__/test_tasks.ts
@@ -13,11 +13,14 @@
 // limitations under the License.
 //
 
-import { Doc } from '@anticrm/core'
+import { Builder, getClass } from '@anticrm/model'
+import { TTask, model as taskModel } from '@anticrm/core-model/src/__tests__/test_tasks'
+import domains from '@anticrm/domains'
 
-/**
- * Indices
- */
-export interface Indices extends Doc {
-  primary: string
+export function model (S: Builder): void {
+  const tclass = getClass(TTask.prototype)
+  tclass.postProcessing.push((model, classifier) => {
+    model.mixinDocument(classifier, domains.mixin.Indices, { primary: 'name' })
+  })
+  taskModel(S)
 }
