+++ v0.4/platform/domains/src/__tests__/domains_tasks.ts
@@ -1,0 +1,4 @@
+export { createTask, doc1, Task, TaskComment, TaskEstimate, taskIds, TaskMixin, TaskWithSecond } from '@anticrm/core/src/__tests__/tasks'
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+export const data = require('./model.json')
